requirements.txt
aniso8601==9.0.1
aocr==0.7.6
appdirs==1.4.4
asgiref==3.7.2
attrs==22.2.0
captcha==0.4
captcha22==1.0.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
Distance==0.1.3
distlib==0.3.8
Django==4.1.4
django-cors-headers==4.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
escape==1.1
et-xmlfile==1.1.0
execnet==1.9.0
filelock==3.13.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-RESTful==0.3.9
fonttools==4.38.0
fuzzy-c-means==1.7.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mock==5.0.1
numpy==1.23.5
opencv-python==4.7.0.68
openpyxl==3.1.2
packaging==21.3
pandas==2.1.3
path==16.6.0
path.py==12.5.0
patsy==0.5.3
Pillow==9.3.0
platformdirs==4.1.0
pluggy==1.0.0
pydantic==1.10.13
pyee==8.2.2
PyJWT==2.8.0
pyparsing==3.0.9
pyppeteer==1.0.2
pytest==7.2.0
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
scipy==1.11.4
six==1.16.0
sqlparse==0.4.3
statsmodels==0.14.0
tabulate==0.8.10
termcolor==2.2.0
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.8.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.25.0
websockets==10.4
Werkzeug==2.2.2
zipp==3.11.0
